pyproject.toml
[project]
name = "rag-document-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.12.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "openai>=2.7.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "wikipedia>=1.4.0",
]
